macros/requirements.txt
requests==2.31.0
gdstk==0.9.59
PyYAML==6.0.1
